+++ database/migrations/2020_05_14_183928_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('product_specifications');
             $table->double('product_price');
             $table->string('product_category');
-            $table->int('category_id');
+            $table->bigInteger('category_id');
 
         });
     }
